refactor: update item and proposal handling to use more descriptive field names; improve item deletion logic based on status

+++ src/services/itemService.js
@@ -1,4 +1,5 @@
 const itemRepository = require("../repository/itemRepository");
+const statusRepository = require("../repository/statusRepository");
 
 async function createItem(itemData, categoryIds, userId, imageFile) {
   const dataForRepo = {
@@ -20,12 +21,23 @@ async function createItem(itemData, categoryIds, userId, imageFile) {
 
 async function updateItem(itemId, itemData, categoryIds, userId, imageFile) {
   const item = await itemRepository.findById(itemId);
-  if (!itemId) {
-    throw new Error("ID do item não fornecido.");
-  }
+
   if (!item) {
     throw new Error("Item não encontrado.");
   }
+
+  const itemStatus = await statusRepository.itemStatusRepository.findByName(
+    "Disponível"
+  );
+
+  if (itemStatus.id != item.statusId) {
+    const err = new Error(
+      "Você só pode deletar itens que estão com o status 'Disponível'."
+    );
+    err.name = "InvalidActionError";
+    throw err;
+  }
+
   if (item.userId !== userId) {
     const err = new Error("Você não tem permissão para atualizar este item.");
     err.name = "UnauthorizedError";
@@ -50,11 +62,16 @@ async function findAllItems(where, page = 1, limit = 10) {
 }
 
 async function deleteItem(itemId, userId) {
-  const item = await itemRepository.findById(itemId);
-
-  if (!itemId) {
-    throw new Error("ID do item não fornecido.");
-  }
+  const item = await itemRepository.findUnique(
+    { id: itemId },
+    {
+      status: {
+        select: {
+          status_name: true,
+        },
+      },
+    }
+  );
 
   if (!item) {
     throw new Error("Item não encontrado.");
@@ -65,6 +82,15 @@ async function deleteItem(itemId, userId) {
     err.name = "UnauthorizedError";
     throw err;
   }
+
+  if (item.status.status_name !== "Disponível") {
+    const err = new Error(
+      "Você só pode deletar itens que estão com o status 'Disponível'."
+    );
+    err.name = "InvalidActionError";
+    throw err;
+  }
+
   return await itemRepository.deleteItem(itemId);
 }
